+++ lib/core/config/constants/api_url.dart
@@ -1,3 +1,5 @@
 class ApiUrl {
   static const String randomReply = 'https://dummyjson.com/quotes/random';
+  static const String dictionaryBase =
+      'https://api.dictionaryapi.dev/api/v2/entries/en/';
 }